migration tested

+++ sandbox.db/src/kz/greetgo/sandbox/db/migration_impl/model/Transaction.java
@@ -1,10 +1,11 @@
 package kz.greetgo.sandbox.db.migration_impl.model;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class Transaction {
   public String type;
-  public double money;
+  public BigDecimal money;
   public String finishedAt;
   public String transactionType;
   public String accountNumber;
